Restructure the code and make two executables
Move the source into a src/ subdirectory.

Add a new executable, driver-standalone, which is intended for one-pass
k-means clustering. It just loads a dataset and clusters it with the
given k and algorithm.  It seems to work but has not been tested much
and is not feature-complete. Some features to add:
    - randomization seeding

The previous executable (which was called "kmeans") is now called
"driver-experiment".

Modify the Makefile a bit to support C++11.